requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
bleach==1.4.2
celery==3.1.23
defusedxml==0.7.1
Django==2.0
django-allauth==0.35.0
django-analytical==2.2.1
django-appconf==1.0.2
django-autocomplete-light==3.1.6
django-celery-email==1.1.5
django-ckeditor==5.5.0
django-cors-headers==2.0.0
django-crispy-forms==1.9.0
django-dbbackup==2.5.0
django-debug-toolbar==1.4
django-distill==1.7
django-embed-video==1.1.2
django-filebrowser==3.10.2
django-filter==0.12.0
django-geojson==2.9.0
django-grappelli==2.12.1
django-haystack==2.5.0
django-honeypot==0.5.0
django-js-asset==1.2.2
django-leaflet==0.18.1
django-post-office==2.0.7
django-taggit==0.23.0
djangorestframework==3.4.0
djangorestframework-gis==0.10.1
drf-haystack==1.5.6
easy-thumbnails==2.4.1
future==0.18.2
html2text==2016.4.2
html5lib==0.9999999
jsonfield==1.0.3
kombu==3.0.35
Markdown==2.6.5
mbutil==0.2.0
mod-wsgi==4.5.2
oauthlib==1.1.1
Pillow==3.1.1
pkg-resources==0.0.0
psycopg2==2.7.5
pysolr==3.5.0
python-dateutil==2.5.0
python-memcached==1.57
python-openid==2.2.5
python3-openid==3.2.0
pytz==2016.3
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
sqlparse==0.1.19
Unipath==1.1
Whoosh==2.7.2
